# preprocS2 v2.2.1 ## Fix - check if (!is.null(out_dir)) at the end of get_cloudmask

+++ R/get_cloudmask.R
@@ -83,6 +83,8 @@ get_cloudmask <- function(collection_path, aoi, iChar, raster_dir, overwrite = F
     }
   }
   if (!is.null(elim)) saveRDS(object = item_collection, file = collection_path)
-  if (dir.exists(out_dir)) unlink(x = out_dir, recursive = T, force = T)
+  if (!is.null(out_dir)){
+    if (dir.exists(out_dir)) unlink(x = out_dir, recursive = T, force = T)
+  }
   return()
 }
